Revert removing `npm config set` when npm publishing
We followed the [documentation](https://github.com/actions/setup-node/blob/main/docs/advanced-usage.md#publish-to-npmjs-and-gpr-with-npm) but got an `ENEEDAUTH` error.

With the `npm config set` approach we got a 404 instead, which suggests a wrong API key.  
However, we know the API key was correct, so the other option was having conflicting ways to configure publishing (`NODE_AUTH_TOKEN` and `npm config set` were both present), or a string interpolation issue (there were no spaces inside the braces).

This commit implements only a single way to configure the auth token.